refactor: modularize scripts
#9
