+++ src/index.html
@@ -100,7 +100,7 @@
 							<button
 								type="button"
 								class="btn btn-primary"
-								id="aviutl-check-version"
+								id="check-core-version"
 							>
 								最新バージョンを確認
 							</button>
@@ -145,7 +145,7 @@
 							<button
 								type="button"
 								class="btn btn-primary w-100"
-								id="aviutl-install"
+								id="install-aviutl"
 							>
 								AviUtlをインストール
 							</button>
@@ -168,7 +168,7 @@
 							<button
 								type="button"
 								class="btn btn-primary w-100"
-								id="exedit-install"
+								id="install-exedit"
 							>
 								拡張編集をインストール
 							</button>
