+++ src/lib/replaceText.js
@@ -1,0 +1,12 @@
+/**
+ * Replaces innerText.
+ *
+ * @param {string} selector - A string of id.
+ * @param {string} text - A text to replace.
+ */
+function replaceText(selector, text) {
+  const element = document.getElementById(selector);
+  if (element) element.innerText = text;
+}
+
+module.exports = replaceText;
